Public Documentation

Documentation for Minuit2.jl public interface.

Index

Minuit2.Minuit

Minuit2.Minuit

Minuit2.FCN

Minuit2.contour

Minuit2.hesse!

Minuit2.matrix

Minuit2.migrad!

Minuit2.minos!

Minuit2.mncontour

Minuit2.mnprofile

Minuit2.profile

Types

This is the list of all types and functions defined for Minuit2

Minuit2.Minuit — Type

Minuit(fcn, x0=(); grad=nothing, error=(), errordef=1.0, names=(), method=:migrad, maxfcn=0, tolerance=0, strategy=1,  kwargs...)

Initialize a Minuit object.

This does not start the minimization or perform any other work yet. Algorithms are started by calling the corresponding methods.

Arguments

fcn::Function : Function to minimize. See notes for details on what kind of functions are accepted.

x0::AbstractArray : Starting values for the minimization. See notes for details on how to set starting values.

grad::Union{Function,Nothing} : If grad is a function, it must be a function that calculates the gradient and returns an iterable object with one entry for each parameter, which is the derivative of fcn for that parameter. If nothing (default), Minuit will compute the gradient numerically.

error::AbstractArray : Starting values for the errors of the parameters. If not provided, Minuit will use 0.1 for all parameters.

errordef::Real : Error definition of the function. Minuit defines parameter errors as the change in parameter value required to change the function value by errordef. Normally, for chisquared fits it is 1, and for negative log likelihood, its value is 0.5. If the user wants instead the 2-sigma errors for chisquared fits, it becomes 4, as Chi2(x+n*sigma) = Chi2(x) + n*n.

names::Sequence{String} : Names of the parameters. If not provided, Minuit will try to extract the names from the function signature.

method::Symbol : The minimization algorithm to use. Possible values are :migrad, :simplex

maxfcn::Int : Maximum number of function calls. If set to 0, Minuit will use a default value.

tolerance::Real : Tolerance for the minimization. If set to 0, Minuit will use a default value.

kwargs : Additional keyword arguments. Starting values for the minimization as keyword arguments. See notes for details on how to set starting values.

Notes

Function to minimize

By default, Minuit assumes that the callable fcn behaves like chi-square function, meaning that the function minimum in repeated identical random experiments is chi-square distributed up to an arbitrary additive constant. This is important for the correct error calculation. If fcn returns a log-likelihood, one should multiply the result with -2 to adapt it. If the function returns the negated log-likelihood, one can alternatively set the attribute errordef to make Minuit calculate errors properly.

Minuit reads the function signature of fcn to detect the number and names of the function parameters. Two kinds of function signatures are understood.

a. Function with positional arguments.

The function has positional arguments, one for each fit parameter. Example:

fcn(a, b, c) =  ...

The parameters a, b, c must accept a real number Minuit automatically detects the parameters names in this case.

b. Function with arguments passed as a single AbstractArray.

The function has a single argument which is an AbstractArray. Example:

function fcn_v(x) =  ...

To use this form, starting values (x0) needs to be passed to Minuit in form of a Tuple or Vector. In some cases, the detection may fail, and will be necessary to use the names keyword to set the parameter names.

Parameter initialization

Initial values for the minimization can be set with positional arguments or via keywords. This is best explained through an example:

fcn(x, y) =  (x - 2)^2 + (y - 3)^2

The following ways of passing starting values are equivalent:

Minuit(fcn, x=1, y=2)
Minuit(fcn, y=2, x=1) # order is irrelevant when keywords are used ...
Minuit(fcn, [1,2])    # ... but here the order matters

Positional arguments can also be used if the function has no signature:

fcn_no_sig(args...) =  ...
Minuit(fcn_no_sig, [1,2])

If the arguments are explicitly named with the names keyword described further below, keywords can be used for initialization:

Minuit(fcn_no_sig, x=1, y=2, names=("x", "y"))  # this also works

If the function accepts a single AbstractVector, then the initial values must be passed as a single array-like object:

fcn_v(x) = return (x[1] - 2) ** 2 + (x[2] - 3) ** 2
Minuit(fcn_v, (1, 2))

Setting the values with keywords is not possible in this case. Minuit deduces the number of parameters from the length of the initialization sequence.

Minuit2.Minuit — Type

Minuit

Minuit object to perform minimization. It keeps track of the function to minimize, the parameters, and the minimization results.

Functions

Minuit2.FCN — Function

FCN(fnc, grad=nothing, arraycall=false, errordef=1.0)

Create a JuliaFcn object from a Julia function fnc and its gradient grad.

Minuit2.contour — Method

contour(m::Minuit, x, y; size=50, bound=2, grid=nothing, subtract_min=false)

Get a 2D contour of the function around the minimum.

It computes the contour via a function scan over two parameters, while keeping all other parameters fixed. The related :meth:mncontour works differently: for each pair of parameter values in the scan, it minimises the function with the respect to all other parameters.

This method is useful to inspect the function near the minimum to detect issues (the contours should look smooth). It is not a confidence region unless the function only has two parameters. Use :meth:mncontour to compute confidence regions.

Arguments

x : First parameter for scan (name or index).

y`: Second parameter for scan (name or index).

size=50 : Number of scanning points per parameter (Default: 50). It can be tuple (nx,ny) as the number of scanning points per parameter. Ignored if grid is set.

bound=2 : Either ((v1min,v1max),(v2min,v2max)) or the number of sigmas to scan symmetrically from minimum.

grid::Tuple{AbstractVector,AbstractVector} : Grid points to scan over. Ifgrid$is set, `size$ and bound are ignored.

subtract_min::Bool=false : Subtract minimum from return values

Returns

Tuple(xv, yv, zv) : Tuple of 1D arrays with the x and y values and a 2D array with the function values.

Minuit2.hesse! — Method

hesse!(m::Minuit, ncall::Int=0)

Run Hesse algorithm to compute asymptotic errors.

The Hesse method estimates the covariance matrix by inverting the matrix of second derivatives (Hesse matrix) at the minimum. To get parameters correlations, you need to use this. The Minos algorithm is another way to estimate parameter uncertainties, see function minos.

Arguments

ncall::Int=0 : Approximate upper limit for the number of calls made by the Hesse algorithm. If set to 0, use the adaptive heuristic from the Minuit2 library.

Notes

The covariance matrix is asymptotically (in large samples) valid. By valid we mean that confidence intervals constructed from the errors contain the true value with a well-known coverage probability (68 % for each interval). In finite samples, this is likely to be true if your cost function looks like a hyperparabola around the minimum.

In practice, the errors very likely have correct coverage if the results from Minos and Hesse methods agree. It is possible to construct artificial functions where this rule is violated, but in practice it should always work.

Minuit2.matrix — Method

matrix(m::Minuit; correlation=false)

Get the covariance matrix of the parameters.

Minuit2.migrad! — Function

migrad!(m::Minuit, strategy=1)

Run Migrad minimization.

Migrad from the Minuit2 library is a robust minimisation algorithm which earned its reputation in 40+ years of almost exclusive usage in high-energy physics. How Migrad works is described in the Minuit paper. It uses first and approximate second derivatives to achieve quadratic convergence near the minimum.

Parameters

m::Minuit : The Minuit object to minimize.

strategy::Int : The minimization strategy. The default value is 1, which is the recommended value for most cases. The value 0 is faster, but less reliable. The value 2 is slower, but more reliable. The value 3 or higher is slower, but even more reliable.

Minuit2.minos! — Method

minos!(m::Minuit, name::String)

Run Minos algorithm to compute asymmetric errors for a single parameter.

The Minos algorithm uses the profile likelihood method to compute (generally asymmetric) confidence intervals. It scans the negative log-likelihood or (equivalently) the least-squares cost function around the minimum to construct a confidence interval.

Arguments

m::Minuit : The Minuit object to minimize.

parameters::AbstractVector{String} : Names of the parameters to compute the Minos errors for.

cl::Number : Confidence level of the interval. If not set, a standard 68 % interval is computed (default). If 0 < cl < 1, the value is interpreted as the confidence level (a probability). For convenience, values cl >= 1 are interpreted as the probability content of a central symmetric interval covering that many standard deviations of a normal distribution. For example, cl=1 is interpreted as 68.3 %, and cl=2 is 84.3 %, and so on. Using values other than 0.68, 0.9, 0.95, 0.99, 1, 2, 3, 4, 5 require the scipy module.

ncall::Int : Limit the number of calls made by Minos. If 0, an adaptive internal heuristic of the Minuit2 library is used (Default: 0).

Notes

Asymptotically (large samples), the Minos interval has a coverage probability equal to the given confidence level. The coverage probability is the probability for the interval to contain the true value in repeated identical experiments.

The interval is invariant to transformations and thus not distorted by parameter limits, unless the limits intersect with the confidence interval. As a rule-of-thumb: when the confidence intervals computed with the Hesse and Minos algorithms differ strongly, the Minos intervals are preferred. Otherwise, Hesse intervals are preferred.

Running Minos is computationally expensive when there are many fit parameters. Effectively, it scans over one parameter in small steps and runs a full minimisation for all other parameters of the cost function for each scan point. This requires many more function evaluations than running the Hesse algorithm.

Minuit2.mncontour — Method

nmcontour(x, y; cl=0.68, size=50, interpolated=0, ncall=0, iterate=5, use_simplex=true)

Get 2D Minos confidence region.

This scans over two parameters and minimises all other free parameters for each scan point. This scan produces a statistical confidence region according to the profile likelihood method with a confidence level cl, which is asymptotically equal to the coverage probability of the confidence region according to Wilks' theorem. Note that 1D projections of the 2D confidence region are larger than 1D Minos intervals computed for the same confidence level. This is not an error, but a consequence of Wilks'theorem.

The calculation is expensive since a numerical minimisation has to be performed at various points.

Arguments

x : Variable name of the first parameter.

y : Variable name of the second parameter.

cl::Real=0.68 : Confidence level of the contour. If not set a standard 68 % contour is computed (default). If 0 < cl < 1, the value is interpreted as the confidence level (a probability). For convenience, values cl >= 1 are interpreted as the probability content of a central symmetric interval covering that many standard deviations of a normal distribution.

size::Int=50 : Number of points on the contour to find. Increasing this makes the contour smoother, but requires more computation time.

interpolated::Int=0 : Number of interpolated points on the contour. If you set this to a value larger than size, cubic spline interpolation is used to generate a smoother curve and the interpolated coordinates are returned. Values smaller than size are ignored. Good results can be obtained with size=20, interpolated=200.

Returns

contour::Vector(Tuple{Float64,Float64}) : Contour points of the form [(x1, y1)...(xn, yn)].

Minuit2.mnprofile — Method

mnprofile(m::Minuit, var; size=30, bound=2, grid=nothing, subtract_min=false, 
               ncall=0, iterate=5, use_simplex=true)

Get Minos profile over a specified interval.

Scans over one parameter and minimises the function with respect to all other parameters for each scan point.

Arguments

var : Parameter to scan over.

size=30 : Number of scanning points. Ignored if grid is set.

bound=2 : If bound is a tuple, (left, right) scanning bound, or the number of sigmas to scan symmetrically around the minimum. Ignored if grid is set.

grid : Parameter values on which to compute the profile. If grid is set, size and bound are ignored.

subtract_min=false : If true, subtract offset so that smallest value is zero.

ncall=0 : Approximate maximum number of calls before minimization will be aborted. If set to 0, use the adaptive heuristic from the Minuit2 library. Note: The limit may be slightly violated, because the condition is checked only after a full iteration of the algorithm, which usually performs several function calls. iterate : int, optional Automatically call Migrad up to N times if convergence was not reached (Default: 5). This simple heuristic makes Migrad converge more often even if the numerical precision of the cost function is low. Setting this to 1 disables the feature. use_simplex: bool, optional If we have to iterate, set this to True to call the Simplex algorithm before each call to Migrad (Default: True). This may improve convergence in pathological cases (which we are in when we have to iterate).

Returns

Tuple(x, y, ok) : Tuple of 1D arrays with the parameter values, function values and booleans whether the fit succeeded or not.

Minuit2.profile — Method

profile(m::Minuit, var; size=100, bound=2, grid=nothing, subtract_min=false)

Calculate 1D cost function profile over a range.

A 1D scan of the cost function around the minimum, useful to inspect the minimum. For a fit with several free parameters this is not the same as the Minos profile computed by mncontour.

Arguments

m::Minuit : The Minuit object to minimize.

var : The parameter to scan over (name or index).

size=100 : Number of scanning points. Ignored if grid is set.

bound=2 : Number of sigmas to scan symmetrically around the minimum. Ignored if grid is set.

grid::AbstractVector : Grid points to scan over. If grid is set, size and bound are ignored.

subtract_min::Bool=false : Subtract minimum from return values.

Returns

Tuple(x, y) : Tuple of 1D arrays with the parameter values and the function values.